+++ types/index.d.ts
@@ -148,6 +148,48 @@ declare namespace Realm {
         flexible: true;
         partitionValue?: never;
         clientReset?: ClientResetConfiguration<ClientResetModeManualOnly>;
+        /**
+         * Optional object to configure the setup of an initial set of flexible
+         * sync subscriptions to be used when opening the Realm. If this is specified,
+         * {@link Realm.open} will not resolve until this set of subscriptions has been
+         * fully synchronized with the server.
+         * 
+         * Example:
+         * ```
+         * const config: Realm.Configuration = {
+         *   sync: {
+         *     user,
+         *     flexible: true,
+         *     initialSubscriptions: {
+         *       update: realm => {
+         *         realm.subscriptions.update(subs => {
+         *           subs.add(realm.objects('Task'));
+         *         })
+         *       }
+         *     }
+         *   },
+         *   // ... rest of config ...
+         * };
+         * const realm = await Realm.open(config);
+         * 
+         * // At this point, the Realm will be open with the data for the initial set
+         * // subscriptions fully synchronised.
+         * ```
+         */
+        initialSubscriptions?: {
+            /**
+             * Callback called with the {@link Realm} instance to allow you to setup the
+             * initial set of subscriptions by calling `realm.subscriptions.update`.
+             * See {@link Realm.App.Sync.SubscriptionSet.update} for more information.
+             */
+            update: (realm: Realm) => void;
+            /**
+             * If `true`, the {@link updateCallback} will be rerun every time the Realm is
+             * opened (e.g. every time a user opens your app), otherwise (by default) it
+             * will only be run if the Realm does not yet exist.
+             */
+            rerunOnStartup?: boolean;
+        };
     }
 
     interface PartitionSyncConfiguration extends BaseSyncConfiguration {